Reset all of organizations state when loading list
This call reassigns the list of all organizations, but doesn't reset the pagination on other included objects. This caused weird data during testing where I'd refresh to test something and the list of organizations would be reloaded, but the celebration items are.

I think the right solution is to reset the pagination of related objects if the parent list of organizations is reloaded for some reason. If we don't a refresh of organizations would nuke all related objects and the related objects couldn't be pulled back in