Implement infinite scrolling on celebrate feed.
Details on known issue in RN:

SectionList has a long standing known issue where `onEndReached` is triggered prematurely when a view is loaded about 50% of the time. This is likely due to an async race condition across the JS bridge. The best known workaround is to implement a state flag to track whether or not the view has been scrolled. This value is defaulted to false on view load so that when the premature onEndReached is called before the view is loaded, the API call is short circuited. Once the view is loaded and ther user interacts with the view the flag is set to true so that when the endThreshold is actually reached the API call is made.

See conversation at: https://github.com/facebook/react-native/issues/16067