Filter peopleByOrgSelector to only return people assigned to me.
the action LOAD_PERSON_DETAILS sticks a person into the allByOrg part of the people state. This is great (I think), but that means if I view another group member who is not assigned to me, it will add them to my assigned people list when going back to the "People" screen.

This also then has to filter the list of orgs, because we may filter "people" on an "org" down to 0 if I view a few members in an organization in which I have no assignments.