Allow GroupsPersonHeader to re-render when state changes
Previously the header had its initial state set, and that was never changed when new props came into the component. The downside to this approach is that when the request that loads the person's details is finished, the shown/hidden buttons are never re-computed.

This does introduce a sight annoyance though in that the view does re-render when the requests finish and the orange "Assign to me" button going away when loading an assigned member is un-sightly. However @5watts and I weighed the pros and cons and decided initially that's better than waiting/blocking for the person download to finish.